pyproject.toml
[project]
name = "InTraj"
version = "0.1.0"
description = "Trajectory prediction zoo"
readme = "README.md"
requires-python = ">=3.10.16,<3.11"
dependencies = [
  "torch==2.4.1",
  "torchvision==0.19.1",
  "torchaudio==2.4.1",
  "torch-cluster @ https://data.pyg.org/whl/torch-2.4.0+cu124/torch_cluster-1.6.3+pt24cu124-cp310-cp310-linux_x86_64.whl",
  "torch-scatter @ https://data.pyg.org/whl/torch-2.4.0+cu124/torch_scatter-2.1.2+pt24cu124-cp310-cp310-linux_x86_64.whl",
  "av2==0.3.5",
  "hydra-core==1.3.2",
  "pytorch-lightning==2.4.0",
  "scipy>=1.10,<2",
  "shapely==2.0.6",
  "torch-geometric==2.6.1",
  "wandb>=0.25.1",
  "waymo-open-dataset-tf-2-12-0>=1.6.5",
  "dotenv>=0.9.9",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]
torchaudio = [{ index = "pytorch-cu124" }]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
